Fix settings view model
