+++ LingoClash/LingoClash/ProfileView/ProfileViewController.swift
@@ -22,18 +22,11 @@ class ProfileViewController: UIViewController {
     override func viewDidLoad() {
         super.viewDidLoad()
         
-        setUpBinders()
         viewModel.refresh()
+        setUpBinders()
     }
     
     func setUpBinders() {
-        
-        viewModel.$error.sink {[weak self] error in
-            if let error = error {
-                self?.showError(error)
-            }
-        }.store(in: &cancellables)
-        
         viewModel.$name.sink {[weak self] name in
             if let name = name {
                 self?.nameLabel.text = name
@@ -58,14 +51,6 @@ class ProfileViewController: UIViewController {
             }
         }.store(in: &cancellables)
         
-        viewModel.$alertContent.sink {[weak self] alertContent in
-            if let alertContent = alertContent {
-                self?.showConfirmAlert(content: alertContent) { _ in
-                    self?.transitionToSplash()
-                }
-            }
-        }.store(in: &cancellables)
-        
         viewModel.$isRefreshing.sink {[weak self] isRefreshing in
             if isRefreshing {
                 self?.showSpinner()
@@ -74,31 +59,4 @@ class ProfileViewController: UIViewController {
             }
         }.store(in: &cancellables)
     }
-    
-    @IBAction func logoutTapped(_ sender: Any) {
-        viewModel.signOut()
-    }
-    
-    func transitionToSplash() {
-        let splashViewController = SplashViewController.instantiateFromAppStoryboard(AppStoryboard.Main)
-        
-        view.window?.rootViewController = splashViewController
-        view.window?.makeKeyAndVisible()
-    }
-    
-    func showError(_ message: String) {
-        // TODO: Perhaps it is better to show as popup
-        Logger.info("Error signing out: \(message)")
-    }
-    
-    override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
-        if let editProfileVC = segue.destination as? EditProfileViewController {
-            editProfileVC.viewModel = self.viewModel
-        } else if let changePasswordVC = segue.destination as? ChangePasswordViewController {
-            changePasswordVC.viewModel = self.viewModel
-        } else if let changeEmailVC = segue.destination as? ChangeEmailViewController {
-            changeEmailVC.viewModel = self.viewModel
-        }
-    }
-    
 }
